metaobject extraction from metaclass
