+++ src/files.cmake
@@ -26,6 +26,7 @@ set(HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/detail/metatype_impl.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/variant.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/metaclass.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/metaobject.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/callable.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/detail/callable_impl.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../include/mox/metadata/metamethod.hpp
@@ -39,6 +40,7 @@ set(SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/metadata_p.h
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/metatype.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/metaclass.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/metadata/metaobject.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/variant.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/callable.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/metadata/signal_p.h
