+++ src/metadata/metaobject.cpp
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2017-2019 bitWelder
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see
+ * <http://www.gnu.org/licenses/>
+ */
+
+#include "metadata_p.h"
+#include <mox/metadata/metaobject.hpp>
+
+namespace mox
+{
+
+MetaObject::MetaObject()
+{
+}
+
+MetaObject::~MetaObject()
+{
+}
+
+const MetaClass* MetaObject::getStaticMetaClass()
+{
+    static ObjectMetaClass<MetaObject> metaClass;
+    return &metaClass;
+}
+
+const MetaClass* MetaObject::getDynamicMetaClass() const
+{
+    return getStaticMetaClass();
+}
+
+} // namespace mox
